+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - New #186: Add `VideoQuality` type.
 - New #186: Add `qualities` field to `Video` type.
 - New #186: Add `firstProfileAudio` field to `ChatFullInfo` type.
+- New #186: Add `UserProfileAudios` type.
 
 ## 0.14 February 7, 2026